autocad_assistance/requirements.txt
numpy>=2.1
pandas==2.2.3
python-telegram-bot>=20.5
chardet==5.2.0
ezdxf>=1.4.2
scipy>=1.14
h11==0.16.0
setuptools>=78.1.1
pyproj>=3.7
simplekml>=1.3.6
